fix: change latitude and longitude types to DOUBLE PRECISION in TimeLog table

+++ src/main/resources/db/migration/V7__create_table_timelog.sql
@@ -2,8 +2,8 @@ CREATE TABLE time_log (
   id UUID PRIMARY KEY,
   date DATE NOT NULL,
   time TIME NOT NULL,
-  latitude DOUBLE,
-  longitude DOUBLE,
+  latitude DOUBLE PRECISION,
+  longitude DOUBLE PRECISION,
   student_id UUID NOT NULL,
   created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
   updated_at TIMESTAMP
